Time: 22 ms (57.89%), Space: 45.2 MB (38.79%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -16,7 +16,7 @@ class Solution {
             double totalHours = 0;
 
             for (int x : piles) {
-                totalHours += Math.ceil((double) x / mid);
+                totalHours += Math.ceil((double) x / (double)mid);
             }
 
             if (totalHours <= h) {
